requirements.txt
openai~=1.12.0
requests~=2.31.0
pillow~=10.2.0